+++ db/migrate/20181009143744_add_recruited_at_to_subjects.rb
@@ -1,0 +1,5 @@
+class AddRecruitedAtToSubjects < ActiveRecord::Migration[5.2]
+  def change
+    add_column :subjects, :recruited_at, :datetime
+  end
+end