+++ db/migrate/20181009151051_add_cover_theme_to_projects.rb
@@ -1,0 +1,5 @@
+class AddCoverThemeToProjects < ActiveRecord::Migration[5.2]
+  def change
+    add_column :projects, :cover_theme, :string
+  end
+end