+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 18.1.0
 
 ### Enhancements
+- **General Changes**
+  - External research projects can be added to research page
 - **Gem Changes**
   - Update to rails 5.2.1
   - Update to pg 1.1.3